Major rewrite to elimnate expandUserTypes and auser.
In the existing type checker, data and alias declarations are
represented by a "user" type constructor that is later replaced
by the actual type. This design was reused in the new type
checker but is now replaced by a more data driven style where
all types are also computed by constraints (as oppsed to the
"convertType" style that tries to compute types upfront).

Example: an alias declared now depends on the availability of
the aliased type. When that is availabel the type of the alias
becomes that of the aliased type.

This is also applied to Syms in SyntaxDefiniitons and all checks
are located in the rule that handles a specific Sym.

Finally, most data structures to represent type info have been
removed: the underlying symbol table is reused.

Example: the constructors for a given ADT are now found via the
symbol table using "useTypeInType".