Configuring the compiler made simpler
The existing situation was that _and_ a TypePalConfig _and_ a
CompilerConfig were needed to configure the typechecker and compiler.

This has been replaced by a single RascalCompilerConfig (see
lang::rascalcore::check::RascalConfig) that contains all configuration
option both for the TypePal specific options and for the
compiler-specific options.

As a result, there is a single place for options, and the signatures of
several functions could be simplified.