Introduce a notion of shallow type checking for IDEs
There is a strong need to speed up typechecking when used in an IDE.
Instead of traversing the complete import/extend graph to spot any
outdated modules, we use a more shallow approach.

If all direct imports/extends of a given top module M have been checked
and are up to date, we stop the exploration of the whole import/extend
graph and check M. Otherwise we fall back to the original exhaustive
exploration of the complete import/extend graph.

The upside of the shallow approach is that it is incredibly much faster
than the original approach. This is very good for the very common
scenario that a user is editing a single module and does not touch its
imports at the same.

The downside of the shallow approach is that more distant changes in the
graph are not automatically discovered, However, if the IDE passes all
changed modules to the typechecker chances are larger (but not 100%)
that remote changes will be discovered. Some further tuning or
adaptation of the shallow approach maybe possible for this.